moves models to models folder
