+++ src/app/shared/models/artist.ts
