+++ src/app/shared/services/api/api.service.ts
@@ -12,8 +12,8 @@ export class ApiService {
   authorization() {
     const client_id = CREDENTIALS.client_id;
     const response_type = 'token';
-    //const redirect_uri = 'http://localhost:4200/stats';
-    const redirect_uri = 'https://spoticheck.me/stats';
+    const redirect_uri = 'http://localhost:4200/stats';
+    // const redirect_uri = 'https://spoticheck.me/stats';
     const scope = 'user-top-read';
     const show_dialog = 'true';
