+++ src/app/stats/stats.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
-import { Track } from '../shared/interfaces/track';
-import { Artist } from '../shared/interfaces/artist';
+import { Track } from '../shared/models/track';
+import { Artist } from '../shared/models/artist';
 import { ApiService } from '../shared/services/api/api.service';
 import { SnackbarService } from "../shared/services/snackbar/snackbar.service";
 // import { ActivatedRoute } from "@angular/router"; // url params
@@ -94,6 +94,7 @@ export class StatsComponent implements OnInit {
   // get the API Data
   async getData(token: String, type: String, timeRange: String = 'medium_term', limit: String = '10', offset: String = '0') {
     let topArray = [];
+    
 
     await this.apiService.getData(token, type, timeRange, limit).toPromise()
       .then((res: any) => {
